# deliveryline-backend/src/main/resources/openapi/openapi.json
{
  "components" : {
    "schemas" : {
...
      "AllowedActions" : {
        "description" : "Backend-derived list of allowed actions for the current state + actor role.",
        "properties" : {
          "actions" : {
            "description" : "Typed action wire values. Primary action first (e.g. approve_spec before reject_spec), then passive views. UI must gracefully ignore unknown values for forward compatibility (UX-DR12, UX-DR6).",
            "example" : [
              "approve_spec",
              "reject_spec",
              "answer_clarification"
            ],
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "versionStamp" : {
            "$ref" : "#/components/schemas/AllowedActionsVersionStamp",
            "description" : "Version stamp the UI echoes back as expectedAllowedActionsVersionStamp on the next mutation so a stale UI surfaces APPROVAL_VERSION_MISMATCH."
          }
        },
        "required" : [
          "actions",
          "versionStamp"
        ],
        "type" : "object"
      },
      "AllowedActionsVersionStamp" : {
        "description" : "Version stamp the UI echoes back on mutations so stale UI surfaces a typed version-mismatch error.",
        "properties" : {
          "currentContextBundleVersion" : {
            "description" : "Context-bundle version of the latest spec's runner execution; null if absent.",
            "example" : 1,
            "format" : "int32",
            "type" : [
              "integer",
              "null"
            ]
          },
          "currentSpecArtifactVersion" : {
            "description" : "Version of the LATEST spec artifact (any approval status); null if none.",
            "example" : 1,
            "format" : "int32",
            "type" : [
              "integer",
              "null"
            ]
          },
          "lastEventId" : {
            "description" : "Public id of the most recent workflow event; null on event-less runs.",
            "example" : "evt_abc123",
            "type" : [
              "string",
              "null"
            ]
          },
          "workflowState" : {
            "description" : "Current workflow state (wire-string form).",
            "example" : "WaitingForSpecApproval",
            "type" : "string"
          }
        },
        "required" : [
          "workflowState"
        ],
        "type" : "object"
      },
...
      "ProblemDetailsResponse" : {
        "description" : "RFC 9457 Problem Details payload with DeliveryLine extension fields.",
        "properties" : {
          "code" : {
            "example" : "RUN_NOT_FOUND",
            "type" : "string"
          },
          "correlationId" : {
            "description" : "Echoed correlation ID when present on the request or MDC.",
            "example" : "3fa85f64-5717-4562-b3fc-2c963f66afa6",
            "type" : [
              "string",
              "null"
            ]
          },
          "detail" : {
            "example" : "Workflow run not found: run_abc123",
            "type" : "string"
          },
          "details" : {
            "anyOf" : [
              {
                "additionalProperties" : { },
                "type" : "object"
              },
              {
                "items" : {
                  "type" : "object"
                },
                "type" : "array"
              }
            ],
            "description" : "Domain-specific extension payload. Validation errors emit an array of field-error objects; other domain errors typically emit an object."
          },
          "instance" : {
            "example" : "/api/v1/workflows/run_abc123/events",
            "type" : "string"
          },
          "retryable" : {
            "example" : false,
            "type" : "boolean"
          },
          "status" : {
            "example" : 404,
            "format" : "int32",
            "type" : "integer"
          },
          "title" : {
            "example" : "Workflow run not found",
            "type" : "string"
          },
          "type" : {
            "example" : "https://deliveryline.local/problems/run-not-found",
            "format" : "uri",
            "type" : "string"
          }
        },
        "required" : [
          "code",
          "detail",
          "instance",
          "retryable",
          "status",
          "title",
          "type"
        ],
        "type" : "object"
      },
...
      "SseEmitter" : {
        "properties" : {
          "timeout" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "WorkflowSummary" : {
        "description" : "Queue/list summary of a workflow run.",
        "properties" : {
          "archivedAt" : {
            "description" : "Soft-hide marker (story 3d-8). Non-null when the run has been archived (hidden from the default queue); null for a live run. The queue renders an archived/hidden badge when present.",
            "format" : "date-time",
            "type" : [
              "string",
              "null"
            ]
          },
          "currentState" : {
            "description" : "Current workflow state.",
            "example" : "WaitingForSpecApproval",
            "type" : "string"
          },
          "escalationMarker" : {
            "description" : "True once specRejectionLoopCount has crossed the configured escalation threshold. Informational — does NOT terminate the workflow (FR13).",
            "example" : false,
            "type" : "boolean"
          },
          "lastEventAt" : {
            "description" : "Timestamp of the most recent event (ISO-8601 UTC).",
            "format" : "date-time",
            "type" : "string"
          },
          "lastEventType" : {
            "description" : "Type of the most recent event.",
            "example" : "workflow.stateChanged",
            "type" : "string"
          },
          "parentRunId" : {
            "description" : "Parent workflow run public id, when this run is a split child.",
            "type" : [
              "string",
              "null"
            ]
          },
          "projectId" : {
            "description" : "Project public id for the run.",
            "example" : "prj_default",
            "type" : [
              "string",
              "null"
            ]
          },
          "projectName" : {
            "description" : "Project display name for the run.",
            "example" : "Default project",
            "type" : [
              "string",
              "null"
            ]
          },
          "projectSlug" : {
            "description" : "Project slug for the run.",
            "example" : "default",
            "type" : [
              "string",
              "null"
            ]
          },
          "specRejectionLoopCount" : {
            "description" : "How many spec rejections this run has accumulated. Increments on every successful rejectSpec call (story 2.10).",
            "example" : 0,
            "format" : "int32",
            "type" : "integer"
          },
          "ticketRef" : {
            "description" : "Linear ticket reference, if linked.",
            "example" : "DEL-1234",
            "type" : "string"
          },
          "ticketTitle" : {
            "description" : "Originating ticket title snapshotted at link time (story 3g-1, FR73). Null for an unlinked or pre-3g run.",
            "example" : "Fix flaky checkout test",
            "type" : [
              "string",
              "null"
            ]
          },
          "workflowRunId" : {
            "description" : "Run public id.",
            "example" : "run_abc123",
            "type" : "string"
          }
        },
        "type" : "object"
      }
    }
  },
  "info" : {
    "description" : "Governed local-first agent delivery workflow API. Phase 1 is localhost-bound (loopback-only) and unauthenticated; the embedded server refuses to bind a non-loopback address without an explicit unsafe override.",
    "title" : "DeliveryLine API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/runner-executions/{rexId}/logs/download" : {
      "get" : {
        "description" : "Returns the runner execution's ALREADY-redacted stdout/stderr as a text/plain attachment (story 3.6 post-hoc redaction is the authoritative guarantee; served verbatim, never re-redacted). Served only over the localhost-only binding to the single local operator; gated by the view_runner_logs allowed-action. The download is recorded as a best-effort audit.logDownloaded event. An unavailable log or gate denial returns 404 RUNNER_EXECUTION_NOT_FOUND.",
        "operationId" : "downloadRunnerLog",
        "parameters" : [
          {
            "description" : "Runner execution public id, e.g. rex_abc123.",
            "example" : "rex_abc123",
            "in" : "path",
            "name" : "rexId",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "Actor role for action gating; view_runner_logs is role-agnostic in the runner-execution states, so the default resolves the gate for any operator.",
            "example" : "workflow_owner",
            "in" : "query",
            "name" : "actorRole",
            "required" : false,
            "schema" : {
              "enum" : [
                "product_reviewer",
                "workflow_owner",
                "developer"
              ],
              "type" : "null"
            }
          },
          {
            "in" : "header",
            "name" : "X-Actor-Identity",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "text/plain" : { }
            },
            "description" : "The redacted runner log as a text/plain attachment."
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetailsResponse"
                }
              }
            },
            "description" : "Malformed runner-execution id (INVALID_ID_PREFIX)."
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetailsResponse"
                }
              }
            },
            "description" : "No such runner execution, no captured/servable log, or view_runner_logs not allowed (RUNNER_EXECUTION_NOT_FOUND)."
          }
        },
        "summary" : "Download a runner execution's redacted log",
        "tags" : [
          "Runner executions"
        ]
      }
    },
...
    "/api/v1/workflows" : {
      "get" : {
        "description" : "Newest-first list of workflow runs for the review queue, optionally filtered by state and project. The projectId query parameter accepts either a project slug or a prj_ public id.",
        "operationId" : "listWorkflows",
        "parameters" : [
          {
            "description" : "Optional current-state filter, e.g. WaitingForSpecApproval.",
            "in" : "query",
            "name" : "state",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "Include soft-hidden (archived) runs. Defaults to false (archived_at IS NULL) — story 3d-8.",
            "in" : "query",
            "name" : "includeArchived",
            "required" : false,
            "schema" : {
              "default" : false,
              "type" : "boolean"
            }
          },
          {
            "description" : "Optional project filter. Accepts a project slug or a prj_ public id.",
            "in" : "query",
            "name" : "projectId",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "Max rows to return (clamped to 1..200).",
            "in" : "query",
            "name" : "limit",
            "required" : false,
            "schema" : {
              "default" : 50,
              "format" : "int32",
              "type" : "integer"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/WorkflowSummary"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Workflow run summaries (direct array, no envelope)."
          }
        },
        "summary" : "List workflow runs",
        "tags" : [
          "Workflows"
        ]
      }
    },
...
    "/api/v1/workflows/{workflowRunId}/allowed-actions" : {
      "get" : {
        "description" : "Returns the typed list of actions valid for the run's current state + actor role, plus a version stamp the UI echoes back on mutations to detect staleness (story 2.14). Read-only and idempotent — no Idempotency-Key required. Forward-compatible: the UI must gracefully hide unknown action values (UX-DR12, UX-DR6).",
        "operationId" : "getAllowedActions",
        "parameters" : [
          {
            "description" : "Run public id, e.g. run_abc123.",
            "example" : "run_abc123",
            "in" : "path",
            "name" : "workflowRunId",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "Actor role for action gating; defaults to product_reviewer when absent. Recognized values are case-sensitive — any other value returns 400 UNKNOWN_ACTOR_ROLE.",
            "example" : "product_reviewer",
            "in" : "query",
            "name" : "actorRole",
            "required" : false,
            "schema" : {
              "enum" : [
                "product_reviewer",
                "workflow_owner",
                "developer"
              ],
              "type" : "null"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AllowedActions"
                }
              }
            },
            "description" : "Allowed actions + version stamp."
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetailsResponse"
                }
              }
            },
            "description" : "Malformed run id (INVALID_ID_PREFIX) or unrecognized actorRole (UNKNOWN_ACTOR_ROLE)."
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetailsResponse"
                }
              }
            },
            "description" : "No such run (RUN_NOT_FOUND)."
          }
        },
        "summary" : "Get backend-derived allowed actions for a workflow run",
        "tags" : [
          "Workflows"
        ]
      }
    },
...
    "/api/v1/workflows/{workflowRunId}/runner-logs/stream" : {
      "get" : {
        "description" : "Server-Sent Events stream of the run's latest runner execution. While the execution is live the container's logs are followed (docker logs --follow) with BEST-EFFORT per-line redaction; once finished, the persisted post-hoc-redacted log (story 3.6) is replayed — that persisted scan is the AUTHORITATIVE redaction guarantee, the live redaction is best-effort only (ADR 0025). Served only over the localhost-only binding to the single local operator; gated by the view_runner_logs allowed-action. Persists nothing and never mutates runner_executions (ADR 0025 D4). Epic 4 story 4.4 ADDED a separate redacted-log ATTACHMENT download (GET /api/v1/runner-executions/{rexId}/logs/download, downloadRunnerLog) — this reverses the earlier 'no separate download surface' note; this SSE viewer remains the live follow. Events: log {stream,line,seq}, status {phase,rex}, end {reason}, error {reason}.",
        "operationId" : "streamRunnerLogs",
        "parameters" : [
          {
            "description" : "Run public id, e.g. run_abc123.",
            "example" : "run_abc123",
            "in" : "path",
            "name" : "workflowRunId",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "Actor role for action gating; defaults to product_reviewer when absent. view_runner_logs is role-agnostic in the runner-execution states, so the default resolves the gate for any operator.",
            "example" : "product_reviewer",
            "in" : "query",
            "name" : "actorRole",
            "required" : false,
            "schema" : {
              "enum" : [
                "product_reviewer",
                "workflow_owner",
                "developer"
              ],
              "type" : "null"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "text/event-stream" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SseEmitter"
                }
              }
            },
            "description" : "An event stream (text/event-stream). When view_runner_logs is not allowed for the run's state, the stream carries a single error+end event instead of log lines."
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetailsResponse"
                }
              }
            },
            "description" : "Malformed run id (INVALID_ID_PREFIX) or unrecognized actorRole."
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetailsResponse"
                }
              }
            },
            "description" : "No such run (RUN_NOT_FOUND)."
          }
        },
        "summary" : "Stream a run's latest runner-execution logs (live + historical)",
        "tags" : [
          "Workflows"
        ]
      }
    },
...
  },
  "servers" : [
    {
      "description" : "Localhost (loopback-only, Phase 1)",
      "url" : "http://127.0.0.1:8080"
    }
  ],
...
}